Ensure member statistics key is decoded
In the StatusController the statistics for the
members on python3 is a byte string which cannot
be jsonified.

This changes so it loops through the statistics
for the member and ensures the key is a string
and not a byte string.

(cherry picked from commit bd2cafadd0961231748cf4b67efd28c6db277b04)

# Conflicts:
#	gnocchi/rest/api.py